Fixed parsing of HotA H3M format, including Seer Huts

+++ lib/mapping/MapFeaturesH3M.h
@@ -17,12 +17,12 @@ enum class EMapFormat : uint8_t;
 struct MapFormatFeaturesH3M
 {
 public:
-	static MapFormatFeaturesH3M find(EMapFormat format, uint8_t hotaVersion);
+	static MapFormatFeaturesH3M find(EMapFormat format, uint32_t hotaVersion);
 	static MapFormatFeaturesH3M getFeaturesROE();
 	static MapFormatFeaturesH3M getFeaturesAB();
 	static MapFormatFeaturesH3M getFeaturesSOD();
 	static MapFormatFeaturesH3M getFeaturesWOG();
-	static MapFormatFeaturesH3M getFeaturesHOTA(uint8_t hotaVersion);
+	static MapFormatFeaturesH3M getFeaturesHOTA(uint32_t hotaVersion);
 
 	MapFormatFeaturesH3M() = default;
